feat!: Send currency object with amounts
Send the entire currency object representation with currency amounts
instead of just the currency code. This allows clients to format the
amount without knowing specific currency codes.
